Sid 203 get info 2 (#35)
* Tx Type precompile bug fixed


* Spotless changes

* OrigchainId information for Originating Transaction added


* Added crosschain id for non crosschain transactions


* Spotless changes


* Acceptance tests changed to use threads instead of processes. Hopefully this will fix memory problems with Circle CI

* Spotless changes

* Web3j libraries updated

+++ acceptance-tests/src/test/java/org/hyperledger/besu/tests/acceptance/crosschain/getinfo/Ctrt2.sol
@@ -23,8 +23,8 @@ contract Ctrt2 is Crosschain, Ctrt2Int {
 
     uint256 public myChainId;
     uint256 public fromChainId;
-    uint32 public myTxType;
-    uint32 public consTxType;
+    uint256 public myTxType;
+    uint256 public consTxType;
     uint256 public coordChainId;
     uint256 public origChainId;
     uint256 public txId;
