+++ acceptance-tests/src/test/java/org/hyperledger/besu/tests/acceptance/crosschain/getinfo/Ctrt3.sol
@@ -18,8 +18,8 @@ import "./Ctrt3Int.sol";
 contract Ctrt3 is Crosschain, Ctrt3Int {
     uint256 public myChainId;
     uint256 public fromChainId;
-    uint32 public myTxType;
-    uint32 public consTxType;
+    uint256 public myTxType;
+    uint256 public consTxType;
     uint256 public coordChainId;
     uint256 public origChainId;
     uint256 public txId;
